pyproject.toml
[project]
name = "erp-skeleton"
version = "0.1.0"
description = "ERP basic skeleton framework for Django"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2,<6.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "markdown>=3.10.2",
]
